Use the new CloudFront Distribution construct
This new construct makes it a bit cleaner to configure a CloudFront
Distribution. Even though the resource is experimental, it seems to be
getting new features (such as the ability to define a managed cache
policy) much faster.

There are some hiccups in migrating to the new construct, with changing
logical IDs. A property to handle keeping the CloudFront Distribution
resource is now included to ease the transition.