Add CloudFront Function for rewrites
Adding this Function allows for using the older format for URLs which
were linked to from several places. Paired with a quick invalidation for
the main page (and possibly other *.html pages) this should take care of
getting back to the older URL format.

The function, unfortunately, has to have linting disabled. This is
because the Functions runtime is extremely limited in syntax. It might
be possible to configure something to transpile it to a syntax but that
feels like significantly more work than just disabling the linter for a
single file.

Reverts 0b1852483d7d4c064bb6b8e43ba908b34a588766
Reverts b49da915830dae75e01e4231ffdfa3199e4c1acb